+++ concordile-spring/concordile-spring-cloud-contract-provider-extension/src/main/java/io/github/concordile/spring/cloud/contract/provider/package-info.java
@@ -15,6 +15,6 @@
  */
 
 @NullMarked
-package io.github.concordile.spring.cloud.contract.producer;
+package io.github.concordile.spring.cloud.contract.provider;
 
 import org.jspecify.annotations.NullMarked;
