+++ concordile-spring/concordile-spring-cloud-contract-provider-extension/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.cloud.contract.verifier.builder.SingleTestGenerator=\
-  io.github.concordile.spring.cloud.contract.producer.ConcordileSingleTestGenerator
+  io.github.concordile.spring.cloud.contract.provider.ConcordileSingleTestGenerator
